requirements.txt
Flask
Flask-SQLAlchemy
qrcode[pil]
Pillow
